Began placeholders for the AST and Parser

+++ main.c
@@ -9,13 +9,21 @@
 
 int main(int argc, char **argv)
 {
+    int passed; // Custom error handling is implemented with this
+    int x;
+    int debug; // Whether or not to print debug messages
+    passed = 1;
+    debug = 0;
+
     printf("OverFlowP\n");
     if(argc < 2){
         printf("Please at least put in 1 argument.\n");
+        passed = 0; // Reports failed
         return(1);
     }
     else if(string_compare(argv[1], "-h") == 1 || string_compare(argv[1], "--help") == 1){
         printf("Commands:\n\t1. help, lists the commands: -h or --help\n");
+        passed = 0; // Reports failed
         return(1);
     }
 
@@ -28,7 +36,10 @@ int main(int argc, char **argv)
 
     while (1) {
         t = tokenize(&pnter); // Get next token
-        printf("Token: type=%d, text='%s'\n", t.type, t.text);
+        if(debug == 1)
+        {
+            printf("Token: type=%d, text='%s'\n", t.type, t.text);
+        }
 
         tokens[token_count] = t;
 
@@ -40,9 +51,22 @@ int main(int argc, char **argv)
         token_count++;
     }
 
-    // Later for AST
+    if(debug == 1)
+    {
+        for(x=0; x < token_count; x++)
+        {
+            printf("%s ", type_converter(tokens[x].type));
+        }
+        printf("\n");
+    }
+    // Later for AST and Parsing
     system("mkdir -p assembling_dir");
 
+
+
     // Assuming Finished with AST and assembling
-    system("rm -rf assembling_dir");
+    if(passed == 1)
+    {
+        system("rm -rf assembling_dir");
+    }
 }
